+++ src/js/handlers/enter.js
@@ -1,6 +1,5 @@
 var ui = require( 'helpers/ui' );
 var fireGAEvent = require( 'helpers/fire-ga-event' );
-var accountLinking = require( 'helpers//account-linking' );
 
 function showNonLDAP( element ) {
   // show social logins + passwordless
@@ -21,8 +20,6 @@ module.exports = function enter( element ) {
   var emailField = document.getElementById( 'field-email' );
   var emailFieldValue = emailField.value.toLowerCase();
   var passwordField = document.getElementById( 'field-password' );
-  var isAccountLinking = accountLinking.isAccountLinking();
-  var qualifiesForLDAPShortcut = false;
   var supportedByRP = form.loginMethods ? form.loginMethods['supportedByRP'] : null;
   var onlyAcceptsLDAP = supportedByRP && supportedByRP.length === 1 && supportedByRP.indexOf( NLX.LDAP_connection_name ) === 0;
   var ENDPOINT = NLX.person_api_domain;
@@ -31,7 +28,7 @@ module.exports = function enter( element ) {
     emailField.focus();
     return;
   }
-  
+
   if ( NLX.features.person_api_lookup ) {
 
     ui.setLockState( element, 'loading' );
@@ -81,5 +78,3 @@ module.exports = function enter( element ) {
     }
   }
 };
-
-